requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.0.7
django-cors-headers==4.4.0
django-extensions==3.2.3
django-filter==24.2
djangorestframework==3.15.2
pika==1.3.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
